Commit 18/12
Important!
- This is a temporary solution of the MyObjectiveFunction as I couldn't
remove all problems of my implementation. It should be okay for tests,
but it isn't good programming (in terms of performance and maybe also of
clarity). Today I won't be able to commit anything else so if possible
wait till friday or at the latest saturday to get a better version of
MyObjectiveFunction.
- Cost.java will probably be modified again as I am not sure if I should
have two separate add methods.
- I noticed that the clone method of MySolution isn't properly
implemented. Please fix it.
- Also I added the parameters alpha, beta and gamma as in the original
MySolution though this is subject to change if we don't need them.