Commit 22/12
Important Note Below!

- I moved around some of the code of the MyObjectiveFunction class to
simplify any changes that might need to be made
- I added a few things to the MySolution class also with the most
important being that the alpha, beta and gamma fields weren't correctly
passed to the solution cloned object.

Important: I have had an error which wouldn't let me run the program at
all to check the efficiency of our program so far. I suspect that the
problem is in one of the following (the lower in the list the least
likely it is that there is a problem):
- Some misunderstanding between all of us
- The clone() method of the MySolution object
- For some reason the evaluate method in MyObjectiveFunction returns
null (this seems impossible, but I don't know what other reason there
could be)