Commit 28/12
Commit which allow the MySearchProgram to split a route into two, and
combine two routes into one.
1) I didn't have time to test them a lot and they give some errors.
Tomorrow, or today, based on perspective I cannot test them until late
afternoon and in the worst case I can only check them from 20.00 -
21.00. So, please try to debug them if you can and also make them more
readable if you can.
2) I removed a method that is no longer necessary from Route.java so as
not to cause confusion.
3) I added a method that allows the removal of a route in
MySolution.java, but it only sets the route in the provided index to
null. So, it doesn't move the possibly non-null routes found below the
removed route into lower indexes, which would be more correct. I have
marked this as a task.
4) For some reason some time routes are added non-sequentially (e.g.
when printing a solution you can get something like R#15 then R#16 and
then R#20) not a very big deal, but it would be better if this didn't
happen. This happened before I changed anything so it's worth looking
into parts of the code that I didn't modify.