Last Commit
I actually got best results with 1(vrptw) and 2(search). Someone should
just run tests using the batch file I will send by email to everyone.
This is especially true for Emmanuel and/or Ivi who will need the data
for the presentation. The presentation should be in formats you can see
in the pdfs provided by the portal so please finish it as quickly as
possible. It needs to be included in what I send tomorrow. This is also
true for an Excel file with the results, so please provide one of those
too. In the meantime I will remove whatever was related to tests and
make ready the jar file, the zip file etc.